warning fix
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk@3453 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ ac/source/src/command.cpp
@@ -1055,7 +1055,7 @@ void currentserver(int i)
                     "acknowledging disconnect",
                     "zombie"
                 };
-                if(curpeer->state>=0 && curpeer->state<sizeof(statenames)/sizeof(statenames[0]))
+                if(curpeer->state>=0 && curpeer->state<int(sizeof(statenames)/sizeof(statenames[0])))
                     s_strcpy(r, statenames[curpeer->state]);
 				break; // 5 == Connected (compare ../enet/include/enet/enet.h +165)
 			}
